moved setup stuff to its own file
